VS: Quote custom command arguments containing semicolons
The `GoogleTest` module needs this to pass the `TEST_EXECUTOR`
definition to its `GoogleTestAddTests.cmake` helper script in
the `POST_BUILD` command.  Previously it worked only if other
characters, such as spaces, caused the argument to be quoted.
This was exposed by running the `RunCMake.GoogleTest` test
in a path without spaces.

Reported-by: Garrett Campbell <gcampbell@microsoft.com>