cmMakefile: invoke callback after command execution
`cmMakefile` has a hook that invokes a callback before each command
execution.  The only user of that hook is `cmCTestScriptHandler`,
which uses it to update the elapsed time since it was started.
The `ctest_sleep` command again updates the elapsed time explicitly.
This explicit update can be saved by invoking the callback after
the command execution instead.