IWYU: Add mapping for jsoncpp version.h
Add an IWYU rule for `version.h` from `jsoncpp`. This mirrors the
existing rules that recommend `value.h` for things from `forwards.h` and
`config.h`, and is needed for IWYU to not complain about use of things
from `version.h`. (This is especially important as we do not directly
wrap `version.h`.)