install: Update CPS export version
Bump the version of CPS we claim to generate from 0.12.0 to 0.13.0,
which is the version we require for import.

The change between 0.12 and 0.13 is the addition of the `prefix`
attribute, and that one of `cps_path` or `prefix` is required. Since our
current implementation always supplies `cps_path`, no other changes are
needed to conform to the new schema version.